changing file structure to have a config firestore
